requirements.txt
# Core dependencies
agno>=0.1.0
typer[all]>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Google Cloud
google-cloud-aiplatform>=1.38.0
google-cloud-discoveryengine>=1.0.0
google-api-core>=2.0.0

# Utilities
python-dateutil>=2.8.0
sqlite3>=3.0.0